dts-admin entity和pipeline'task接口和测试类编写

+++ dts-admin-dao/src/main/java/com/utopia/data/transfer/admin/dao/mapper/RegionBeanMapper.java
@@ -9,7 +9,7 @@ import org.apache.ibatis.annotations.Param;
 public interface RegionBeanMapper extends BaseDao {
     int deleteByExample(RegionBeanDal example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(RegionBean record);
